card-game-visualizer/server/requirements.txt
aiohttp==3.9.3
python-socketio==5.11.1
gymnasium==0.29.1
numpy==1.24.3
sb3-contrib==2.2.1
stable-baselines3==2.2.1
